Extract ConnectorTableCredentialsVisitor to top level
